requirements.txt
amqp==5.2.0
celery==5.4.0
Django==5.0.7
django-autoslug==1.9.9
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-filter==24.3
django-guardian==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
dj_database_url==2.2.0
drf-yasg==1.21.7
Faker==27.0.0
gunicorn==23.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.8
setuptools==72.2.0
weasyprint==62.3
openai==1.39.0
psycopg==3.1.9  # Add this for psycopg
